requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0

# scraper dependancies
curl_cffi
beautifulsoup4
lxml
numpy
requests
urllib3
python-dotenv
pandas
boto3 # for AWS dynamodb connection (optional)
supabase